% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_dampener.R
\name{calculate_dampener}
\alias{calculate_dampener}
\title{Calculate the dampener}
\usage{
calculate_dampener(to = Sys.Date())
}
\arguments{
\item{to}{The final date for which the index is calculated}
}
\value{
xts value of length 1 with the size of the dampener effect
}
\description{
Calculate the dampener
}
\examples{
calculate_dampener()
}

