% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plot_components}
\alias{plot_components}
\title{Plot the components of the dampener}
\usage{
plot_components(s2_index)
}
\arguments{
\item{s2_index}{- The index used for the calculation of the dampener}
}
\value{
ggplot of the components of the dampener
}
\description{
Plot the components of the dampener
}
\details{
The dampener effect consists of 4 components:
1. The last value of the index, 2. the average value of the index,
3. a constant to subtract from the relative distance between the last and the average value of the index.
}
\examples{
s2_index = load_data()
plot_components(s2_index)
}

