% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solvency2EquityDampener.R
\docType{package}
\name{solvency2EquityDampener-package}
\alias{solvency2EquityDampener-package}
\title{Calculate the equity dampener for the standard formula of Solvency II.}
\description{
Calculate the equity dampener for the standard formula of Solvency II.
}
\details{
\tabular{ll}{
Package: \tab solvency2EquityDampener\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-07-21\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

\code{\link{calculate_dampener}} calculates the dampener for a given end date \code{to}.

For this, it downloads the equity indizes, as specified in ITS ... as proposed by EIOPA in CP 14/058,
calculates the aggregate index, and applies the formula for the dampener as specified in Article 172
of the Delegated Acts:

SA = max(-0.1, min(0.1({(CI-AI)}{AI} - 0.08))), with  CI - current value of index
and AI - average value of index over last 36 months.
}
\examples{
calculate_dampener("2014-12-31")
#                  [,1]
# 2014-12-31 0.03168399
}
\author{
Owe Jessen \email{post@owejessen.de}
}
\references{
\url{http://ec.europa.eu/internal_market/insurance/docs/solvency/solvency2/delegated/141010-delegated-act-solvency-2_en.pdf}
\url{https://eiopa.europa.eu/Publications/Reports/EIOPA-BoS-15-120_Final_report_ITS_Equity_dampener.pdf}
}
\keyword{II}
\keyword{Solvency}
\keyword{package,}

