% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapper_official_dampener.R
\name{wrapper_official_dampener}
\alias{wrapper_official_dampener}
\title{Calculate the dampener based on official index}
\usage{
wrapper_official_dampener(to = Sys.Date())
}
\arguments{
\item{to}{The final date for which the index is calculated}
}
\value{
xts value of length 1 with the size of the dampener effect
}
\description{
Calculate the dampener based on official index
}
\examples{
wrapper_official_dampener()
}

